refactor: 💡 extract asker info components and minor cleanups
